Fixed issues with spline_bp(). Remove degree arg?
The main idea behind this method is to emulate v-slope, jm, and orr methods, meaning it's a piecewise linear regression with straight lines. Changing the degree argument at this time won't give correct values, so it might be worth removing it so users don't get weird results without knowing it.