Dealing with lm fitting to no data
There are corner cases where the Orr and similar methods have the intersection point in the first minute of the data, when then usually gets removed by the front trim argument. Since you get an error if you try to fit an lm to a data frame with no rows, that broke the functions. This code should fix that.