Fix indeterminant findings bugs
Occasionally, the regression models would have NA values in the coefficients. Other times, the p values in the loop_res would have NA results. Both caused the function to break. Both of these scenarios imply that the model does not fit the data well, so the result can be classified as indeterminant.